requirements.txt
pyyaml>=6.0
sqlalchemy>=2.0
chromadb>=0.4.0
requests>=2.31.0
pydantic>=2.0
python-dotenv>=1.0
aiohttp>=3.9.0
